+++ types/protocol-mapping.d.ts
@@ -368,6 +368,23 @@ export namespace ProtocolMapping {
          */
         'Network.directTCPSocketClosed': [Protocol.Network.DirectTCPSocketClosedEvent];
         /**
+         * Fired when data is sent to tcp direct socket stream.
+         */
+        'Network.directTCPSocketChunkSent': [Protocol.Network.DirectTCPSocketChunkSentEvent];
+        /**
+         * Fired when data is received from tcp direct socket stream.
+         */
+        'Network.directTCPSocketChunkReceived': [Protocol.Network.DirectTCPSocketChunkReceivedEvent];
+        /**
+         * Fired when there is an error
+         * when writing to tcp direct socket stream.
+         * For example, if user writes illegal type like string
+         * instead of ArrayBuffer or ArrayBufferView.
+         * There's no reporting for reading, because
+         * we cannot know errors on the other side.
+         */
+        'Network.directTCPSocketChunkError': [Protocol.Network.DirectTCPSocketChunkErrorEvent];
+        /**
          * Fired when additional information about a requestWillBeSent event is available from the
          * network stack. Not every requestWillBeSent event will have an additional
          * requestWillBeSentExtraInfo fired for it, and there is no guarantee whether requestWillBeSent